Changes to make cogent compatible with python2.4

+++ cogent/maths/stats/ks.py
@@ -108,7 +108,8 @@ def psmirnov2x(statistic, least, most):
     q = floor(statistic * most * least - 1e-7) / (least * most)
     u_vector = zeros(most+1, 'd')
     for j in range(most+1):
-        u_vector[j] = [1,0][j / most > q]
+        #SUPPORT2425
+        u_vector[j] = [1,0][int(j / most > q)]
     for i in range(1,least+1):
         w = i / (i+most)
         if i/least > q:
